backend/requirements.txt
fastapi==0.95.2
starlette==0.27.0
pydantic==1.10.7
uvicorn==0.22.0
python-dotenv==1.0.1
